+++ granulate_utils/linux/cgroups_client/cpu_cgroup.py
@@ -3,7 +3,7 @@
 # Licensed under the AGPL3 License. See LICENSE.md in the project root for license information.
 #
 
-from granulate_utils.linux.cgroups.base_cgroup import BaseCgroup
+from granulate_utils.linux.cgroups_client.base_cgroup import BaseCgroup
 
 
 class CpuCgroup(BaseCgroup):
